refact!: New `Kirby\Panel\Areas` class
BREAKING CHANGE: `KIrby\Panel\Panel` returns `Kirby\Panel\Areas` object instead of an array.
BREAKING CHANGE: Moved `KIrby\Panel\Panel::area()` to Kirby\Panel\Areas::area()`
BREAKING CHANGE: Moved `KIrby\Panel\Panel::buttons()` to Kirby\Panel\Areas::buttons()`
