+++ src/Panel/Menu.php
@@ -73,7 +73,7 @@ class Menu
 			// merge area definition (e.g. from config)
 			// with global area definition
 			if (is_array($area) === true) {
-				$area = Panel::area($id, [
+				$area = Areas::area($id, [
 					...$this->areas[$id] ?? [],
 					'menu' => true,
 					...$area
