+++ tests/Panel/PanelTest.php
@@ -50,111 +50,10 @@ class PanelTest extends TestCase
 		unset($_SERVER['SERVER_SOFTWARE']);
 	}
 
-	public function testArea(): void
-	{
-		// defaults
-		$result = Panel::area('test', []);
-		$expected = [
-			'id' => 'test',
-			'label' => 'test',
-			'breadcrumb' => [],
-			'breadcrumbLabel' => 'test',
-			'title' => 'test',
-			'menu' => false,
-			'link' => 'test',
-			'search' => null
-		];
-
-		$this->assertSame($expected, $result);
-	}
-
 	public function testAreas(): void
 	{
-		// unauthenticated / uninstalled
 		$areas = Panel::areas();
-
-		$this->assertArrayHasKey('installation', $areas);
-		$this->assertCount(1, $areas);
-
-		// create the first admin
-		$this->app = $this->app->clone([
-			'users' => [
-				[
-					'email' => 'test@getkirby.com',
-					'role'  => 'admin'
-				]
-			]
-		]);
-
-		// unauthenticated / installed
-		$areas = Panel::areas();
-
-		$this->assertArrayHasKey('login', $areas);
-		$this->assertArrayHasKey('logout', $areas);
-		$this->assertCount(2, $areas);
-
-		// simulate a logged in user
-		$this->app->impersonate('test@getkirby.com');
-
-		// authenticated
-		$areas = Panel::areas();
-
-		$this->assertArrayHasKey('search', $areas);
-		$this->assertArrayHasKey('site', $areas);
-		$this->assertArrayHasKey('system', $areas);
-		$this->assertArrayHasKey('users', $areas);
-		$this->assertArrayHasKey('account', $areas);
-		$this->assertArrayHasKey('logout', $areas);
-		$this->assertArrayHasKey('lab', $areas);
-		$this->assertCount(7, $areas);
-
-		// authenticated with plugins
-		$app = $this->app->clone([
-			'areas' => [
-				'todos' => fn () => []
-			]
-		]);
-
-		$app->impersonate('test@getkirby.com');
-
-		$areas = Panel::areas();
-
-		$this->assertArrayHasKey('todos', $areas);
-		$this->assertCount(8, $areas);
-	}
-
-	public function testButtons(): void
-	{
-		$this->app = $this->app->clone([
-			'users' => [
-				[
-					'email' => 'test@getkirby.com',
-					'role'  => 'admin'
-				]
-			]
-		]);
-
-		$this->app->impersonate('test@getkirby.com');
-		$core = Panel::buttons();
-
-		// add custom buttons
-		$this->app = $this->app->clone([
-			'areas' => [
-				'test' => fn () => [
-					'buttons' => [
-						'a' => ['component' => 'test-a'],
-						'b' => ['component' => 'test-b']
-					]
-				]
-			]
-		]);
-
-		$this->app->impersonate('test@getkirby.com');
-		$withCustoms = Panel::buttons();
-
-		$this->assertSame(2, count($withCustoms) - count($core));
-		$this->assertSame(['component' => 'test-b'], array_pop($withCustoms));
-		$this->assertSame(['component' => 'test-a'], array_pop($withCustoms));
+		$this->assertInstanceOf(Areas::class, $areas);
 	}
 
 	public function testFirewallWithoutUser(): void
